Ticket #1435: Add PubSub support on Windows (UDP) part 2

+++ src/Common/helpers_platform_dep/windows/p_sopc_sockets.c
@@ -177,7 +177,7 @@ SOPC_Socket_Address* SOPC_Socket_GetPeerAddress(SOPC_Socket sock)
         if (0 == res)
         {
             result->address.ai_family = sockAddrStorage->ss_family;
-            result->address.ai_addrlen = sockAddrStorageLen;
+            result->address.ai_addrlen = (size_t) sockAddrStorageLen;
             result->address.ai_addr = (struct sockaddr*) sockAddrStorage;
         }
     }
